use postresql as data source

+++ db/migrate/20180318223509_create_projects.rb
@@ -1,6 +1,9 @@
 class CreateProjects < ActiveRecord::Migration[5.1]
   def change
-    drop_table :projects
+
+    if ActiveRecord::Base.connection.table_exists? :projects
+      drop_table :projects
+    end
     create_table :projects do |t|
       t.string :name
